+++ src/components/chat/insurance-list/InsuranceList.tsx
@@ -1,4 +1,3 @@
-import { ExtraBuilding } from 'components/chat/tabs/InsuranceTab'
 import { LinkRow } from 'components/shared'
 import { WideModal } from 'components/shared/modals/WideModal'
 import PaginatorList from 'components/shared/paginator-list/PaginatorList'
@@ -19,6 +18,31 @@ import {
 const DateTypeEnum = {
   terminationDate: 1,
   activationDate: 2,
+}
+
+export enum ExtraBuildingType {
+  ATTEFALL = 'Attefallshus',
+  SAUNA = 'Bastu',
+  BOATHOUSE = 'Båthus',
+  CARPORT = 'Carport',
+  GUESTHOUSE = 'Gästhus',
+  FRIGGEBOD = 'Friggebod',
+  STOREHOUSE = 'Förråd',
+  GARAGE = 'Garage',
+  BARN = 'Lada',
+  GAZEBO = 'Lusthus',
+  SHED = 'Skjul',
+  OUTHOUSE = 'Uthus',
+  GREENHOUSE = 'Växthus',
+  OTHER = 'Annat',
+}
+
+export interface ExtraBuilding {
+  id: string
+  type: ExtraBuildingType
+  area: number
+  hasWaterConnected: boolean
+  displayName?: string
 }
 
 export default class InsuranceList extends React.Component {
